refactor: split controller into different modules
